fix: clean instances
